getSessionId method and remove .env from source control

+++ src/app.py
@@ -17,7 +17,7 @@ app = Flask(__name__)
 
 
 def validate_secret():
-    env_secret = os.getenv('SECRET')
+    env_secret = config.get('SECRET')
     request_secret = request.headers.get('X-Secret')
 
     if not env_secret:
@@ -25,6 +25,32 @@ def validate_secret():
 
     return request_secret == env_secret
 
+
+@app.route("/getSessionId", methods=['POST'])
+def test():
+    if not validate_secret():
+        return 'Invalid secret', 401
+    
+    body = request.get_json(silent=True)
+
+    if not body:
+        return 'Invalid request or missing body', 400
+
+    ig_username = body.get('igUsername')
+    ig_password = body.get('igPassword')
+
+    if not ig_username or not ig_password:
+      return 'Invalid request, missing parameters', 400
+
+    c = Client()
+    success = c.login(ig_username, ig_password)
+
+    if not success:
+        return 'Invalid ig credentials', 400
+    
+    return {
+        'sessionId': c.session_id
+    }
 
 @app.route("/uploadIGTVVideo", methods=['POST'])
 def upload_igtv_video():
@@ -36,26 +62,26 @@ def upload_igtv_video():
     if not body:
         return 'Invalid request or missing body', 400
 
-    ig_username = body.get('igUsername')
-    ig_password = body.get('igPassword')
+    session_id = body.get('sessionId')
     video_url = body.get('videoURL')
     title = body.get('title')
     caption = body.get('caption')
     thumbnail_url = body.get('thumbnailURL')
 
-    if not ig_username or not ig_password or not video_url \
-            or not title or not caption or not thumbnail_url:
+    if not session_id or not video_url \
+            or not title or not caption:
         return 'Invalid request, missing parameters', 400
 
     try:
         video_path = f'upload/{time.time()}.mp4'
-        thumbnail_path = f'upload/{time.time()}.jpg'
+        thumbnail_path = f'upload/{time.time()}.jpg' if thumbnail_url else None
 
         urllib.request.urlretrieve(video_url, video_path)
-        urllib.request.urlretrieve(thumbnail_url, thumbnail_path)
+        if thumbnail_url:
+          urllib.request.urlretrieve(thumbnail_url, thumbnail_path)
 
         cl = Client()
-        success = cl.login(ig_username, ig_password)
+        success = cl.login_by_sessionid(session_id)
 
         if not success:
             print('Invalid ig credentials, login failed')
@@ -66,13 +92,14 @@ def upload_igtv_video():
             Path(video_path),
             title,
             caption,
-            Path(thumbnail_path)
+            thumbnail=Path(thumbnail_path) if thumbnail_path else None
         )
 
         print(f'Media uploaded successfully!, id: {media.id}')
 
         os.remove(video_path)
-        os.remove(thumbnail_path)
+        if thumbnail_path:
+          os.remove(thumbnail_path)
 
         return media.id
     except Exception as e:
